+++ changelog.md
@@ -2,6 +2,7 @@
 * Made the TagParser accept codec-form CompoundTag (NBT) Objects.
 * Fixed a crash that could happen if negative luck was encountered.
 * Added new Supporter Wings (Armored, Blaze, Bone, Cloud, Fairy, Mechanical, and Monarch).
+* Added CachedObject, which allows for caching of nbt-encoded itemstack objects.
 
 ## 7.1.7
 * Fixed a network encode error in ItemAdapter.CODEC.